course-gen-service/requirements.txt
openai>=1.0.0
pika>=1.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
